pyproject.toml
[tool.poetry]
name = "animes-api"
version = "0.1.0"
description = ""
authors = ["Damego <danyabatueff@gmail.com>"]
readme = "README.md"
packages = [{include = "animes_api"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
asyncpg = "^0.27.0"
pydantic = {extras = ["email"], version = "^1.10.6"}
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
cryptography = "^39.0.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
